+++ smtlib-backends-z3/CHANGELOG.md
@@ -1,3 +1,6 @@
+# v0.3-alpha
+- make test-suite compatible with `smtlib-backends-0.3`
+
 # v0.2
 - make test-suite compatible with `smtlib-backends-0.2`
 - add usage examples in the test-suite